Overload Encodable operators
